Separation sub navigators into modules; Adding types for navigators
