+++ package.json
@@ -25,6 +25,7 @@
     "react-native-safe-area-context": "0.6.0",
     "react-native-screens": "2.0.0-alpha.12",
     "react-native-web": "~0.11.7",
+    "react-navigation": "^4.2.2",
     "react-navigation-header-buttons": "^3.0.5",
     "styled-components": "^5.0.1"
   },
